labs: ignore locks for debug + fix log in lab0

+++ labs/lab0_test.sh
@@ -33,10 +33,13 @@ check() {
 
 # lock lab
 lock() {
+    if [ -n "$DEBUG" ]; then
+        return 0
+    fi
     files=(/tmp/lab*)
     if [ -e "${files[0]}" ]; then
         base_name=$(basename "${files[0]}")
-        echo "$base_name was already created, please run: sudo ./$base_name delete"
+        log "$base_name was already created, please run: sudo ./$base_name delete" "ERROR"
         exit 1
     fi
     touch "/tmp/$ME"
@@ -45,6 +48,9 @@ lock() {
 
 # unlock lab
 unlock() {
+    if [ -n "$DEBUG" ]; then
+        return 0
+    fi
     rm -f "/tmp/$ME" || true
 }
